refactor(triggers): extract trigger functions into a separate module
- Removed trigger-related functions from `git_utils.py` and `response_agent.py`.
- Created a new file `triggers.py` to house these functions.
- Updated references in `response_agent.py` to use functions from the `triggers` module.
