+++ src/response_agent.py
@@ -13,7 +13,6 @@ from github.Repository import Repository
 from git_utils import (
     get_github_client,
     get_repository,
-    has_blech_bot_tag,
     write_issue_response,
     get_issue_details,
     clone_repository,
@@ -29,59 +28,12 @@ from agents import (
     gemerate_prompt,
     generate_prompt,
 )
-
+import triggers
 
 ############################################################
 # Trigger checks
 ############################################################
 
-def has_generate_edit_command_trigger(issue: Issue) -> bool:
-    """
-    Check if the issue comments contain the trigger for generate_edit_command
-
-    Args:
-        issue: The GitHub issue to check
-
-    Returns:
-        True if the trigger phrase is found in any comment
-    """
-    comments = get_issue_comments(issue)
-    return any("[ generate_edit_command ]" in comment.body for comment in comments)
-
-def has_bot_response(issue: Issue) -> bool:
-    """
-    Check if there is already a bot response on the issue
-
-    Args:
-        issue: The GitHub issue to check
-
-    Returns:
-        True if there is a comment from blech_bot
-    """
-    comments = get_issue_comments(issue)
-    return any("generated by blech_bot" in comment.body for comment in comments)
-
-
-def has_user_feedback(issue: Issue) -> bool:
-    """
-    Check if there is user feedback after the latest bot response
-
-    Args:
-        issue: The GitHub issue to check
-
-    Returns:
-        True if there is a non-bot comment after the latest bot comment
-    """
-    comments = get_issue_comments(issue)
-
-    # Find the latest bot comment
-    latest_bot_idx = -1
-    for i, comment in enumerate(comments):
-        if "generated by blech_bot" in comment.body:
-            latest_bot_idx = i
-
-    # Check if there are any comments after the latest bot comment
-    return latest_bot_idx >= 0 and latest_bot_idx < len(comments) - 1
 
 ############################################################
 # Response patterns 
@@ -283,9 +235,9 @@ def check_triggers(issue: Issue) -> str:
     Returns:
         The trigger phrase found in the issue
     """
-    if has_user_feedback(issue):
+    if triggers.has_user_feedback(issue):
         return "feedback"
-    elif has_generate_edit_command_trigger(issue):
+    elif triggers.has_generate_edit_command_trigger(issue):
         return "generate_edit_command"
     else:
         return "new_response"
@@ -323,11 +275,11 @@ def process_issue(
     try:
         # Check if issue has blech_bot tag or blech_bot in title, and no existing response
         if not ignore_checks:
-            has_bot_mention = has_blech_bot_tag(
+            has_bot_mention = triggers.has_blech_bot_tag(
                 issue) or "[ blech_bot ]" in issue.title.lower()
             if not has_bot_mention:
                 return False, "Issue does not have blech_bot tag or mention in title"
-            if has_bot_response(issue) and not has_user_feedback(issue):
+            if has_bot_response(issue) and not triggers.has_user_feedback(issue):
                 return False, "Issue already has a bot response without feedback from user"
 
         # Generate and post response
